refactor: move buffer and stream helpers to `@crawlee/core`
`isStream`, `isBuffer`, `toBuffer` and `weightedAvg` were `@ignore`-d helpers in `@crawlee/utils` whose only consumer is `@crawlee/core`, so they now live there as package-private internals.
