+++ packages/utils/src/internals/general.ts
@@ -15,23 +15,6 @@ export const URL_WITH_COMMAS_REGEX =
     /https?:\/\/(www\.)?([\p{L}0-9]|[\p{L}0-9][-\p{L}0-9@:%._+~#=]{0,254}[\p{L}0-9])\.[a-z]{2,63}(:\d{1,5})?(\/[-\p{L}0-9@:%_+,.~#?&/=()'*]*)?/giu;
 
 /**
- * Computes a weighted average of an array of numbers, complemented by an array of weights.
- * @ignore
- */
-export function weightedAvg(arrValues: number[], arrWeights: number[]): number {
-    const result = arrValues
-        .map((value, i) => {
-            const weight = arrWeights[i];
-            const sum = value * weight;
-
-            return [sum, weight];
-        })
-        .reduce((p, c) => [p[0] + c[0], p[1] + c[1]], [0, 0]);
-
-    return result[0] / result[1];
-}
-
-/**
  * Returns a `Promise` that resolves after a specific period of time. This is useful to implement waiting
  * in your code, e.g. to prevent overloading of target website or to avoid bot detection.
  *
@@ -99,56 +82,4 @@ export function expandShadowRoots(document: Document): string {
     replaceShadowDomsWithHtml(document.body);
 
     return document.documentElement.outerHTML;
-}
-
-/**
- * Checks if the given value is a Node.js Stream or a Web API ReadableStream.
- * @ignore
- */
-export function isStream(value: unknown): value is NodeJS.ReadableStream | ReadableStream {
-    if (typeof value !== 'object' || value === null) {
-        return false;
-    }
-
-    // A Node.js Readable is both pipeable and async-iterable; a Web ReadableStream exposes pipeTo.
-    // Requiring async-iterability for the `pipe` branch rejects plain `{ pipe }` ducks that would
-    // otherwise blow up later in the storage backends' drain loop with a cryptic TypeError.
-    const isNodeStream =
-        typeof (value as any).pipe === 'function' && typeof (value as any)[Symbol.asyncIterator] === 'function';
-    const isWebStream = typeof (value as any).pipeTo === 'function';
-
-    return isNodeStream || isWebStream;
-}
-
-/**
- * Checks if the given value is a Node.js Buffer, ArrayBuffer, or TypedArray.
- * @ignore
- */
-export function isBuffer(value: unknown): value is Buffer | ArrayBuffer | ArrayBufferView {
-    return (
-        value != null &&
-        typeof value === 'object' &&
-        (Buffer.isBuffer(value) ||
-            value instanceof ArrayBuffer ||
-            ArrayBuffer.isView(value) ||
-            (value as any).constructor?.name === 'Buffer')
-    );
-}
-
-/**
- * Converts a byte-like value (Buffer, ArrayBuffer, or any typed-array / DataView) into a Buffer over
- * the exact same bytes, honoring `byteOffset` / `byteLength` for views. Existing Buffers are returned
- * as-is. Used by storage backends, which persist raw bytes regardless of the input's concrete shape.
- * @ignore
- */
-export function toBuffer(value: Buffer | ArrayBuffer | ArrayBufferView): Buffer {
-    if (Buffer.isBuffer(value)) {
-        return value;
-    }
-
-    if (value instanceof ArrayBuffer) {
-        return Buffer.from(value);
-    }
-
-    return Buffer.from(value.buffer, value.byteOffset, value.byteLength);
 }
